# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Tugas UKK Peminjaman Alat",
        "description": "Dokumentasi API Sistem Peminjaman Alat",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/chat/users": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Ambil daftar user untuk chat",
                "operationId": "0d6df1c8ee89c4a61f52ff364aff6fdb",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chat/messages/{userId}": {
            "get": {
                "tags": [
                    "Chat"
                ],
                "summary": "Ambil pesan chat dengan user tertentu",
                "operationId": "853fd09664a8b91823b35657d150c895",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil pesan chat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chat/messages": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Kirim pesan chat",
                "operationId": "741403c2b0321371cd3b2f6f2da53af5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "receiver_id"
                                ],
                                "properties": {
                                    "receiver_id": {
                                        "type": "integer"
                                    },
                                    "message": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pesan berhasil dikirim"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chat/messages/{id}": {
            "delete": {
                "tags": [
                    "Chat"
                ],
                "summary": "Hapus pesan chat",
                "operationId": "3fe217d3309659afa5ec20c8312d6dd2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pesan berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/chat/messages/{userId}/read": {
            "post": {
                "tags": [
                    "Chat"
                ],
                "summary": "Tandai pesan sebagai sudah dibaca",
                "operationId": "6c5e1128f424d59b84545e14d11649d0",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pesan berhasil ditandai sebagai dibaca"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/laporan": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Ambil data laporan peminjaman",
                "operationId": "50a73a12a1dd9b19550a543e8aa1e9b6",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data laporan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/laporan/export/excel": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Export laporan ke Excel",
                "operationId": "2fd16a9a78cdf95796311ea202bf3ca6",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendownload file Excel"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/laporan/export/pdf": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Export laporan ke PDF",
                "operationId": "61ac4de4cbe97d3dc4a3c1600605ad32",
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendownload file PDF"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}